+++ packages/gatsby/src/query/queue.js
@@ -22,10 +22,6 @@ const createBuildQueue = () => {
 }
 
 const createDevelopQueue = getRunner => {
-  let queue
-  const processing = new Set()
-  const waiting = new Map()
-
   const queueOptions = {
     ...createBaseOptions(),
     priority: (job, cb) => {
@@ -37,17 +33,6 @@ const createDevelopQueue = getRunner => {
     },
     merge: (oldTask, newTask, cb) => {
       cb(null, newTask)
-    },
-    // Filter out new query jobs if that query is already running.
-    // When the query finshes, it checks the waiting map and pushes
-    // another job to make sure all the user changes are captured.
-    filter: (job, cb) => {
-      if (processing.has(job.id)) {
-        waiting.set(job.id, job)
-        cb(`already running`)
-      } else {
-        cb(null, job)
-      }
     },
   }
 
@@ -66,33 +51,20 @@ const createDevelopQueue = getRunner => {
           })
         }
 
-        processing.delete(queryJob.id)
-        if (waiting.has(queryJob.id)) {
-          queue.push(waiting.get(queryJob.id))
-          waiting.delete(queryJob.id)
-        }
         callback(null, result)
       },
       error => callback(error)
     )
   }
 
-  queue = new Queue(handler, queueOptions)
-  return queue
+  return new Queue(handler, queueOptions)
 }
-
-const pushJob = (queue, job) =>
-  new Promise((resolve, reject) =>
-    queue
-      .push(job)
-      .on(`finish`, resolve)
-      .on(`failed`, reject)
-  )
 
 /**
  * Returns a promise that pushes jobs onto queue and resolves onces
  * they're all finished processing (or rejects if one or more jobs
  * fail)
+ * Note: queue is reused in develop so make sure to thoroughly cleanup hooks
  */
 const processBatch = async (queue, jobs, activity) => {
   let numJobs = jobs.length
@@ -100,15 +72,33 @@ const processBatch = async (queue, jobs, activity) => {
     return Promise.resolve()
   }
 
-  const runningJobs = jobs.map(job =>
-    pushJob(queue, job).then(v => {
-      if (activity.tick) {
-        activity.tick()
-      }
-      return v
-    })
-  )
-  return Promise.all(runningJobs)
+  return new Promise((resolve, reject) => {
+    if (activity.tick) {
+      queue.on(`task_finish`, () => activity.tick())
+    }
+
+    const gc = () => {
+      queue.removeAllListeners(`task_failed`)
+      queue.removeAllListeners(`drain`)
+      queue.removeAllListeners(`task_finish`)
+      queue = null
+    }
+
+    queue
+      // Note: the first arg is the path, the second the error
+      .on(`task_failed`, (...err) => {
+        gc()
+        reject(err)
+      })
+      // Note: `drain` fires when all tasks _finish_
+      //       `empty` fires when queue is empty (but tasks are still running)
+      .on(`drain`, () => {
+        gc()
+        resolve()
+      })
+
+    jobs.forEach(job => queue.push(job))
+  })
 }
 
 module.exports = {