+++ packages/gatsby/src/schema/types/built-in-types.js
@@ -1,0 +1,80 @@
+const { parse } = require(`graphql`)
+
+const fileType = `
+  type File implements Node @infer {
+    sourceInstanceName: String!
+    absolutePath: String!
+    relativePath: String!
+    extension: String!
+    size: Int!
+    prettySize: String!
+    modifiedTime: Date! @dateformat
+    accessTime: Date! @dateformat
+    changeTime: Date! @dateformat
+    birthTime: Date! @dateformat
+    root: String!
+    dir: String!
+    base: String!
+    ext: String!
+    name: String!
+    relativeDirectory: String!
+    dev: Int!
+    mode: Int!
+    nlink: Int!
+    uid: Int!
+    gid: Int!
+    rdev: Int!
+    ino: Float!
+    atimeMs: Float!
+    mtimeMs: Float!
+    ctimeMs: Float!
+    atime: Date! @dateformat
+    mtime: Date! @dateformat
+    ctime: Date! @dateformat
+    birthtime: Date @deprecated(reason: "Use \`birthTime\` instead")
+    birthtimeMs: Float @deprecated(reason: "Use \`birthTime\` instead")
+  }
+`
+
+const directoryType = `
+  type Directory implements Node @infer {
+    sourceInstanceName: String!
+    absolutePath: String!
+    relativePath: String!
+    extension: String!
+    size: Int!
+    prettySize: String!
+    modifiedTime: Date! @dateformat
+    accessTime: Date! @dateformat
+    changeTime: Date! @dateformat
+    birthTime: Date! @dateformat
+    root: String!
+    dir: String!
+    base: String!
+    ext: String!
+    name: String!
+    relativeDirectory: String!
+    dev: Int!
+    mode: Int!
+    nlink: Int!
+    uid: Int!
+    gid: Int!
+    rdev: Int!
+    ino: Float!
+    atimeMs: Float!
+    mtimeMs: Float!
+    ctimeMs: Float!
+    atime: Date! @dateformat
+    mtime: Date! @dateformat
+    ctime: Date! @dateformat
+    birthtime: Date @deprecated(reason: "Use \`birthTime\` instead")
+    birthtimeMs: Float @deprecated(reason: "Use \`birthTime\` instead")
+  }
+`
+
+const builtInTypeDefinitions = () =>
+  [fileType, directoryType].map(type => parse(type))
+
+module.exports = {
+  builtInTypeDefinitions,
+}