+++ packages/gatsby-source-filesystem/CHANGELOG.md
@@ -3,6 +3,32 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.46](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-filesystem@2.1.45...gatsby-source-filesystem@2.1.46) (2020-01-09)
+
+**Note:** Version bump only for package gatsby-source-filesystem
+
+## [2.1.45](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-filesystem@2.1.43...gatsby-source-filesystem@2.1.45) (2020-01-09)
+
+**Note:** Version bump only for package gatsby-source-filesystem
+
+## [2.1.44](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-filesystem@2.1.43...gatsby-source-filesystem@2.1.44) (2020-01-09)
+
+**Note:** Version bump only for package gatsby-source-filesystem
+
+## [2.1.43](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-filesystem@2.1.42...gatsby-source-filesystem@2.1.43) (2019-12-20)
+
+### Bug Fixes
+
+- **gatsby-source-filesystem:** fix add/change text ([#20188](https://github.com/gatsbyjs/gatsby/issues/20188)) ([0fda1e6](https://github.com/gatsbyjs/gatsby/commit/0fda1e6))
+
+## [2.1.42](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-filesystem@2.1.40...gatsby-source-filesystem@2.1.42) (2019-12-10)
+
+**Note:** Version bump only for package gatsby-source-filesystem
+
+## [2.1.41](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-filesystem@2.1.40...gatsby-source-filesystem@2.1.41) (2019-12-10)
+
+**Note:** Version bump only for package gatsby-source-filesystem
+
 ## [2.1.40](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-filesystem@2.1.39...gatsby-source-filesystem@2.1.40) (2019-12-02)
 
 **Note:** Version bump only for package gatsby-source-filesystem